refactor: 공통되는 reducer 따로 파일 분리
